requirements.txt
faiss-cpu==1.7.4
numpy==1.24.3
openai==0.27.8
pandas==1.5.3
boto3==1.28.23